push math, fix audit comment

+++ smart-contracts/contracts/cl-vault/src/state.rs
@@ -51,8 +51,8 @@ pub enum SwapDirection {
 #[cw_serde]
 pub struct ModifyRangeState {
     // pre-withdraw state items
-    pub lower_tick: i128,
-    pub upper_tick: i128,
+    pub lower_tick: i64,
+    pub upper_tick: i64,
     // pre-deposit state items
     pub new_range_position_ids: Vec<u64>,
 }
